Reviewed-on: https://source.greenglobal.vn/tools-dntn/crm-api/pulls/289

+++ src/crm/admission-register/src/services/StudentService.php
@@ -49,10 +49,16 @@ class StudentService
     public static function deleteStudent(string $id)
     {
         $url = env('SSO_URL') . '/api/students/' . $id;
-        
         $token = request()->bearerToken();
+
+        $getStudent = Http::withToken($token)->get($url);
+        $dataStudent = json_decode($getStudent->body());
+
+        if (!empty($dataStudent->student) && !is_null($dataStudent->student->classId)) {
+            throw new HttpException($getStudent->status(), 'Học sinh đã được phân bổ không được hủy đăng ký');
+        }
         $response = Http::withToken($token)->delete($url);
-        
+
         if ($response->failed()) {
             $message = 'Có lỗi từ api CLOVER';
